+++ packages/to-json-schema/src/convertAction.ts
@@ -3,15 +3,30 @@ import type * as v from 'valibot';
 import type { ConversionConfig } from './type.ts';
 import { handleError } from './utils/index.ts';
 
-// TODO: Add support for more actions (for example all regex-based actions)
-
 /**
  * Action type.
  */
 type Action =
   | v.Base64Action<string, v.ErrorMessage<v.Base64Issue<string>> | undefined>
+  | v.BicAction<string, v.ErrorMessage<v.BicIssue<string>> | undefined>
+  | v.Cuid2Action<string, v.ErrorMessage<v.Cuid2Issue<string>> | undefined>
+  | v.DecimalAction<string, v.ErrorMessage<v.DecimalIssue<string>> | undefined>
   | v.DescriptionAction<unknown, string>
+  | v.DigitsAction<string, v.ErrorMessage<v.DigitsIssue<string>> | undefined>
   | v.EmailAction<string, v.ErrorMessage<v.EmailIssue<string>> | undefined>
+  | v.EmojiAction<string, v.ErrorMessage<v.EmojiIssue<string>> | undefined>
+  | v.EmptyAction<
+      v.LengthInput,
+      v.ErrorMessage<v.EmptyIssue<v.LengthInput>> | undefined
+    >
+  | v.HexadecimalAction<
+      string,
+      v.ErrorMessage<v.HexadecimalIssue<string>> | undefined
+    >
+  | v.HexColorAction<
+      string,
+      v.ErrorMessage<v.HexColorIssue<string>> | undefined
+    >
   | v.IntegerAction<number, v.ErrorMessage<v.IntegerIssue<number>> | undefined>
   | v.Ipv4Action<string, v.ErrorMessage<v.Ipv4Issue<string>> | undefined>
   | v.Ipv6Action<string, v.ErrorMessage<v.Ipv6Issue<string>> | undefined>
@@ -29,10 +44,6 @@ type Action =
       v.LengthInput,
       number,
       v.ErrorMessage<v.LengthIssue<v.LengthInput, number>> | undefined
-    >
-  | v.NonEmptyAction<
-      v.LengthInput,
-      v.ErrorMessage<v.NonEmptyIssue<v.LengthInput>> | undefined
     >
   | v.MaxLengthAction<
       v.LengthInput,
@@ -59,8 +70,15 @@ type Action =
       number,
       v.ErrorMessage<v.MultipleOfIssue<number, number>> | undefined
     >
+  | v.NanoIDAction<string, v.ErrorMessage<v.NanoIDIssue<string>> | undefined>
+  | v.NonEmptyAction<
+      v.LengthInput,
+      v.ErrorMessage<v.NonEmptyIssue<v.LengthInput>> | undefined
+    >
+  | v.OctalAction<string, v.ErrorMessage<v.OctalIssue<string>> | undefined>
   | v.RegexAction<string, v.ErrorMessage<v.RegexIssue<string>> | undefined>
   | v.TitleAction<unknown, string>
+  | v.UlidAction<string, v.ErrorMessage<v.UlidIssue<string>> | undefined>
   | v.UrlAction<string, v.ErrorMessage<v.UrlIssue<string>> | undefined>
   | v.UuidAction<string, v.ErrorMessage<v.UuidIssue<string>> | undefined>
   | v.ValueAction<
@@ -89,6 +107,20 @@ export function convertAction(
       break;
     }
 
+    case 'bic':
+    case 'cuid2':
+    case 'decimal':
+    case 'digits':
+    case 'emoji':
+    case 'hexadecimal':
+    case 'hex_color':
+    case 'nanoid':
+    case 'octal':
+    case 'ulid': {
+      jsonSchema.pattern = valibotAction.requirement.source;
+      break;
+    }
+
     case 'description': {
       jsonSchema.description = valibotAction.description;
       break;
@@ -96,6 +128,21 @@ export function convertAction(
 
     case 'email': {
       jsonSchema.format = 'email';
+      break;
+    }
+
+    case 'empty': {
+      if (jsonSchema.type === 'array') {
+        jsonSchema.maxItems = 0;
+      } else {
+        if (jsonSchema.type !== 'string') {
+          handleError(
+            `The "${valibotAction.type}" action is not supported on type "${jsonSchema.type}".`,
+            config
+          );
+        }
+        jsonSchema.maxLength = 0;
+      }
       break;
     }